Add TemperatureSensor
Adds TemperatureSensor class to heaters.py without using it

TemperatureSensor is a wrapper for hardware sensors that offers smoothed temperature and smoothed temperature derivative. This provides consistent  consistent and useful measurements for other modules to use.

This commit should be safe to keep because it adds code that is not used.

Signed-off-by: Gareth Farrington <gareth@waves.ky>