Attempt to resolve spiky readings
Saw spikes in the load cell graph. This could be due to reading the sensors too fast. Re-add the timed delays but use irq_poll() version of the delay.

Also don't send the host any data when the pin of a sensor was found still high low. This should indicate a miss-read of the sensor.