Various Small Fixes
There are cases where a read can take a long time (when the step queue is flushed to the MCU) and thats OK for the HX71x so long as the read doesn't take longer than the time period between samples.

Stop logging exceptions on shutdown when the sensor tried to finish measurements but the serial port is already shut down. This lead to the sensor being blamed to for the shutdown in the log when it was not he root cause.

Stop logging HX71x samples, which polluted the logs

Fix pending flag logic to match the official PR